# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Smart Blog",
        "description": "Aplicación web que genera blogs con ayuda de ChatGPT",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/blog": {
            "get": {
                "tags": [
                    "Blog"
                ],
                "summary": "Listado de los blogs vigentes del usuario",
                "operationId": "544a6146a7525f7c997ecf0fd015c79c",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "user_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "status_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "topic": {
                                                        "type": "string",
                                                        "example": "SpaceX"
                                                    },
                                                    "blog": {
                                                        "type": "string",
                                                        "example": "Space Exploration Technologies Corp., conocida como SpaceX, es una empresa estadounidense de fabricación aeroespacial y de servicios de transporte espacial con sede en Hawthorne. Fue fundada en 2002 por Elon Musk con el objetivo de reducir los costes de viajar al espacio para facilitar la colonización de Marte.​​​"
                                                    },
                                                    "expires_at": {
                                                        "type": "timestamp",
                                                        "example": "2023-02-23T00:09:16.000000Z"
                                                    },
                                                    "created_at": {
                                                        "type": "timestamp",
                                                        "example": "2023-02-23T00:09:16.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "timestamp",
                                                        "example": "2023-02-23T12:33:45.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "SERVER ERROR",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ocurrió un problema al ingresar a esta vista. Contacta a soporte técnico"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/blog/filter": {
            "post": {
                "tags": [
                    "Blog"
                ],
                "summary": "Filtra los blogs por tema",
                "operationId": "51e5bfe7e1edfadf5a8172cb5875d1ba",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "filter": {
                                                        "type": "string",
                                                        "example": "SpaceX"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "SERVER ERROR",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ocurrió un problema al filtrar los blogs. Contacta a soporte técnico"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/blog/store": {
            "post": {
                "tags": [
                    "Blog"
                ],
                "summary": "Crear un blog",
                "operationId": "b2be6851e4274300b9d975b0fb737016",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "topic": {
                                                        "type": "string",
                                                        "example": "SpaceX"
                                                    },
                                                    "chatgpt_response": {
                                                        "type": "string",
                                                        "example": "Space Exploration Technologies Corp., conocida como SpaceX, es una empresa estadounidense de fabricación aeroespacial y de servicios de transporte espacial con sede en Hawthorne. Fue fundada en 2002 por Elon Musk con el objetivo de reducir los costes de viajar al espacio para facilitar la colonización de Marte.​​​"
                                                    },
                                                    "blog_date": {
                                                        "type": "timestamp",
                                                        "example": "2023-02-23T00:09:16.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "SERVER ERROR",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ocurrió un problema al guardar tu post. Contacta a soporte técnico"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}